refactor: adjust name module utils to shared
