+++ android/processor-shared-code/gradle.properties
@@ -14,5 +14,5 @@
 # limitations under the License.
 #
 
-POM_ARTIFACT_ID=android-processor-utils
-POM_NAME=android-processor-utils
+POM_ARTIFACT_ID=android-processor-shared-code
+POM_NAME=android-processor-shared-code
