+++ android/processor/src/main/java/br/com/zup/beagle/android/compiler/BeagleAnnotationProcessor.kt
@@ -21,9 +21,7 @@ import br.com.zup.beagle.android.annotation.RegisterController
 import br.com.zup.beagle.android.annotation.RegisterValidator
 import br.com.zup.beagle.annotation.RegisterAction
 import br.com.zup.beagle.annotation.RegisterWidget
-import br.com.zup.beagle.compiler.BEAGLE_CONFIG
-import br.com.zup.beagle.compiler.error
-import br.com.zup.beagle.compiler.implementsInterface
+import br.com.zup.beagle.compiler.shared.error
 import com.google.auto.service.AutoService
 import net.ltgt.gradle.incap.IncrementalAnnotationProcessor
 import net.ltgt.gradle.incap.IncrementalAnnotationProcessorType
