+++ android/processor/src/main/java/br/com/zup/beagle/android/compiler/KotlinPoetExtensions.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package br.com.zup.beagle.compiler
+package br.com.zup.beagle.android.compiler
 
 import javax.lang.model.element.TypeElement
 import javax.lang.model.type.DeclaredType
