+++ android/processor/src/main/java/br/com/zup/beagle/android/compiler/RegisterControllerProcessor.kt
@@ -17,10 +17,7 @@
 package br.com.zup.beagle.android.compiler
 
 import br.com.zup.beagle.android.annotation.RegisterController
-import br.com.zup.beagle.compiler.BEAGLE_ACTIVITY
-import br.com.zup.beagle.compiler.CONTROLLER_REFERENCE
-import br.com.zup.beagle.compiler.DEFAULT_BEAGLE_ACTIVITY
-import br.com.zup.beagle.compiler.error
+import br.com.zup.beagle.compiler.shared.error
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FileSpec
