+++ android/processor/src/main/java/br/com/zup/beagle/android/compiler/RegisterWidgetProcessor.kt
@@ -16,10 +16,10 @@
 
 package br.com.zup.beagle.android.compiler
 
-import br.com.zup.beagle.compiler.BEAGLE_CORE_WIDGET
-import br.com.zup.beagle.compiler.BeagleSetupRegisteredWidgetGenerator
-import br.com.zup.beagle.compiler.REGISTERED_WIDGETS
-import br.com.zup.beagle.compiler.error
+import br.com.zup.beagle.compiler.shared.BeagleSetupRegisteredWidgetGenerator
+import br.com.zup.beagle.compiler.shared.REGISTERED_WIDGETS
+import br.com.zup.beagle.compiler.shared.BEAGLE_CORE_WIDGET
+import br.com.zup.beagle.compiler.shared.error
 import com.squareup.kotlinpoet.AnnotationSpec
 import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
