+++ android/processor/src/main/java/br/com/zup/beagle/android/compiler/beaglesetupmanage/TypeElementImplementationManager.kt
@@ -17,14 +17,14 @@
 package br.com.zup.beagle.android.compiler.beaglesetupmanage
 
 import br.com.zup.beagle.android.compiler.PropertySpecifications
-import br.com.zup.beagle.compiler.FORM_LOCAL_ACTION_HANDLER
-import br.com.zup.beagle.compiler.DEEP_LINK_HANDLER
-import br.com.zup.beagle.compiler.HTTP_CLIENT_HANDLER
-import br.com.zup.beagle.compiler.STORE_HANDLER
-import br.com.zup.beagle.compiler.URL_BUILDER_HANDLER
-import br.com.zup.beagle.compiler.BEAGLE_LOGGER
-import br.com.zup.beagle.compiler.BEAGLE_IMAGE_DOWNLOADER
-import br.com.zup.beagle.compiler.implementsInterface
+import br.com.zup.beagle.android.compiler.FORM_LOCAL_ACTION_HANDLER
+import br.com.zup.beagle.android.compiler.DEEP_LINK_HANDLER
+import br.com.zup.beagle.android.compiler.HTTP_CLIENT_HANDLER
+import br.com.zup.beagle.android.compiler.STORE_HANDLER
+import br.com.zup.beagle.android.compiler.URL_BUILDER_HANDLER
+import br.com.zup.beagle.android.compiler.BEAGLE_LOGGER
+import br.com.zup.beagle.android.compiler.BEAGLE_IMAGE_DOWNLOADER
+import br.com.zup.beagle.android.compiler.implementsInterface
 import javax.annotation.processing.ProcessingEnvironment
 import javax.lang.model.element.TypeElement
